+++ changelog.md
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 * Switched to using Ketting
+* Added the 'registry:new' command, allowing you to add new event types to the
+  registry.
 
 
 0.0.1 (2023-02-27)